Progress on taking API key input
