+++ src/App.tsx
@@ -11,39 +11,6 @@ import DetailedPage from './pages/DetailedPage';
 import HomePage from './pages/homePage';
 import BasicPage from './pages/BasicPage';
 
-import key from './pages/homePage';
-
-import OpenAI from "openai";
-
-const openai = new OpenAI({
-  ///organization:,
-  ///apiKey: key,
-  dangerouslyAllowBrowser: true,
-});
-
-async function main() {
-  const response = await openai.chat.completions.create({
-    model: "gpt-4-turbo",
-    messages: [
-      {
-        "role": "system",
-        "content": "You will tell me what career I should pursue based on my interests."
-      },
-      {
-        "role": "user",
-        "content": "I like math, computers, and logic."
-      }
-    ],
-    temperature: 0.8,
-    max_tokens: 64,
-    top_p: 1,
-  });
-
-  console.log(response.choices[0].message.content);
-}
-
-main();
-
 function App() {
    
   return (
